refactor(extension): extract sidepanel hooks and utilities from SidepanelApp
Split ~400 lines of inline logic out of SidepanelApp.tsx into focused,
reusable modules:

- useDragAndDropZone – drag-and-drop event handling
- useScrollPinToBottom – scroll-to-bottom pin/button state
- useSidepanelContextMenu – runtime context-menu command listener
- useTabContextWatcher – tab activation/update watcher
- useTitleGeneration – session title generation lifecycle

Extracted pure utilities:
- utils/dropPayload.ts – DataTransfer parsing helpers
- utils/pendingContextCommand.ts – PendingSidepanelContextCommand type/guards
- utils/titleGeneration.ts – LLM title generation helpers
- isComposingEnterEvent moved to utils/dom.ts
